refactor testutils, exclude from build, modify build/test triggers
